Move Optical Flow from aux to pipeline/operators/optical_flow
Order includes lexicographically
