+++ dali/CMakeLists.txt
@@ -33,8 +33,6 @@ add_subdirectory(pipeline)
 add_subdirectory(util)
 add_subdirectory(plugin)
 add_subdirectory(c_api)
-add_subdirectory(aux)
-
 # Collect source files for dali
 file(GLOB tmp *.cc)
 set(DALI_SRCS ${DALI_SRCS} ${tmp})
