+++ dali/pipeline/operators/CMakeLists.txt
@@ -19,14 +19,15 @@ add_subdirectory(decoder)
 add_subdirectory(detection)
 add_subdirectory(displacement)
 add_subdirectory(fused)
+add_subdirectory(geometric)
+add_subdirectory(optical_flow)
+add_subdirectory(paste)
 add_subdirectory(reader)
 add_subdirectory(resize)
 add_subdirectory(sequence)
 add_subdirectory(support)
-add_subdirectory(util)
-add_subdirectory(paste)
-add_subdirectory(geometric)
 add_subdirectory(transpose)
+add_subdirectory(util)
 
 # Get all the source files and dump test files
 file(GLOB tmp *.cc *.cu)
