+++ dali/pipeline/operators/optical_flow/optical_flow_adapter/optical_flow_adapter_test.cc
@@ -14,8 +14,9 @@
 
 #include <gtest/gtest.h>
 #include <memory>
-#include "dali/aux/optical_flow/optical_flow_stub.h"
+
 #include "dali/kernels/backend_tags.h"
+#include "dali/pipeline/operators/optical_flow/optical_flow_adapter/optical_flow_stub.h"
 #include "dali/util/cuda_utils.h"
 
 namespace dali {
